requirements.txt
azure-functions==1.24.0
MarkupSafe==3.0.3
pyodbc==5.3.0
python-dotenv==1.2.1
Werkzeug==3.1.3
